Deps: SMBJ 0.11.5 -> 0.14.0
The pin was for anonymous authentication breaking in 0.12; 0.14.0 was
tried against an impacket SMB2 guest share (anonymous connect, listing,
read and write) and worked. SMBJ now pulls the jdk18on Bouncy Castle,
which is excluded like everywhere else, and File.write() returns a long.
The pin comment says how to go back if a NAS disagrees.

+++ app/src/main/java/me/zhanghai/android/files/provider/smb/client/FileByteChannel.kt
@@ -66,7 +66,7 @@ class FileByteChannel(
         } catch (e: SMBRuntimeException) {
             throw e.toIOException()
         }
-        source.position(sourcePosition + bytesWritten)
+        source.position(sourcePosition + bytesWritten.toInt())
     }
 
     @Throws(IOException::class)
